fix(replace-sqlx): apply clippy review findings
Delete dead-code duplication in src/classification/types.rs (74 lines):
PatternEntry, DualThreshold, NegativePatternConfig, CategoryConfig and
their default_* helpers were moved to src/config/types.rs but the
originals were left behind, generating 9 dead_code warnings. Mirrors
9c92b458 fix from the code-structure-reorg branch.

Drop the unnecessary `as i32` cast in sql_backend.rs:704 — `i` from
`for i in 0..3` is already `i32`.

Result: cargo clippy --all-targets → No issues found (was 10 warnings);
cargo test → 363 passed.

Touched: src/classification/types.rs, src/persistence/sql_backend.rs.

+++ src/persistence/sql_backend.rs
@@ -701,7 +701,7 @@ mod tests {
                 status: "ok".to_string(),
                 category: Some(format!("CAT_{}", i)),
                 upstream_model: Some("model".to_string()),
-                duration_ms: Some(10 * i as i32),
+                duration_ms: Some(10 * i),
                 prompt_snippet: format!("snippet {}", i),
                 prompt_char_count: Some(50),
                 created_at: chrono::Utc::now(),
